Adding in more helper scripts.
The link database can now be managed as a single file in the _data
directory, and the linkref plugin can reach into it and pull data out
based on UID. This way, we can create a "database" of resources, and
both 1) render it out as a page from Liquid, and 2) reach into it from
the plugin for rendering links throughout the site.

Did this so as to consolidate links. Managing links in a site has always
bothered me, and this seems like a good/quick solution.

Need to start working in branches.