Renaming, Reorg
I was working on pagenav, but... I don't understand the disconnect, at
this moment, between the version of uswds-jekyll that I have on my
machine and the documentation in Github. That is, things in the docs
actually break my site.

So.

I've done some renaming for consistency. The include lets us turn a slug
into a long form describing the indicator of a page.