Adding in a Joel Test page for the SOs.
This is part of two pieces of writing that operationalize
recommendations from the State Field Guide/Derisking Guide.

We recommend SOs ask technical questions, but that doesn't mean that
they know what technical questions to ask. The Joel Test is venerable,
and covers a lot of the basics. This writeup expands the Joel Test in
the specific context of civic projects that SOs might be overseeing with
their states.