Adding in _courses collection.
This collection provides structure for the "course" separate from the
rubric. Now, a course is generally a single "overview", which collects
up links to the lessons in the course. Content is linked to from the
"rubric" collection.

Some changes, and more to come in this content. It can now be cleaned up
a fair bit. By adding summaries to the front matter of the rubric pages,
these overviews can become very template-y. Course two has this as an
example, course one needs cleanup, course three is already following
this structure.