Handling multiple lessons on one ID
If I have multiple lessons for (say) id1.3, they all need to appear on
the course overview page. This might not be the right way to do this...
perhaps the course overview should not be rubric dimensions, but true
UIDs to the content.

This, in truth, is because the data model for the rubric lessons is
broken. We should consider fixing it, so each rubric lesson page has a
UID.