Improved menu.
This version of the menu differs from the version in uswds-jekyll. I
personally think it might be a bit cleaner/simpler, but it is also
(possibly) doing less work.

Regardless, it builds a semantically/functionally correct menu, which is
what matters. And, it does not rely on the existence of the uswds module
for Jekyll, which cannot work on GH pages.