Fixed alert/link tags in federalist version.
Both alert and link were ruby modules.

Now they are liquid includes.

This means that they work on GH Pages, and therefore work on Federalist.
However, this is more portable, because it is a pure liquid solution.

This involves touches in all the files that use the references. Those
changes are made here.