Adding prompts to CF conversation
Adding in the audio and prompts for the PO/CF conversation.

The prompts are essentially going to be the same for all of the audio
pieces. We may, therefore, want to invest a bit more in the structure of
the listening exercises. Right now, they are prompted to take notes on
two listens, focusing on their own reflections and practice.

The conversation provides a structure to make sure everyone is heard,
and to capture the learning of the group for sharing with other SOs. So,
the structure is *ok*, but it might be good to spend a bit more time
thinking about what would make it as "high impact" as possible.