This cleanup removes plugin deps.
This merges the simplify branch back into staging.

This now renders 5-7x faster, and makes static content out of things
that were Jekyll modules. In short, this is now able to be deployed to
GH Pages (if we wanted to), and will definitely render/work in any
deployment environment we are presented with.