Numbering, useless metadata
Removed useless metadata from the top of pages. These were removed in
"simplify," but not in staging. The merge process brought it back.

I'm increasingly confident there was a better way to do the merge. Or,
the branches should not have been allowed to be so far apart. That was
on me.

Fixed how the rubric linking to pages works, so now it will catch one
row having multiple lessons, and visa-versa.

We have one lesson left to link in.