Adding in Admin section
This section will have pieces-parts regarding the running of the course.
The first addition is the "Learning Cohorts" piece that was written as a
GDoc for CMS, and then converted to the GH Pages wiki by Amy. I've
copy-pasted it into the site, because... well, for transparency, the
admin and running of the course should probably go with the course?

We can pull it if we want to. But, why not make the package complete and
stand-alone?