Normalizing lesson layouts.
Based on noodling over a ticket this afternoon, it is clear that lessons
could have a more normalized layout.

Given that we have a blurb at the top of every lesson (the "overview"),
and that overview is already used on the overview page, it became
apparent that this could be used as the intro to the lesson, too.

Now, the "lesson" layout includes the title of the lesson, the
overview, a table of contents, and the rubric dimension. This way, the
lessons themselves are just the lesson. This will help with various
things looking forward.

I've made the requisite changes to the supporting scripts, and have
moved course one over to the "lesson" layout. I'll do the other courses
next, as it really simplifies things.