Adding CSS class for screen reader support
The rubric page now does the following:

1. The DIV-based version of the rubric is now invisible to screen
   readers.
2. On the same page, the entire *text-based* rubric is rendered.
2a. The text-based rubric is invisible to people who are *not* using a
screen reader.

This way, when a screen reader comes upon this content, what the reader
gets is content structured as plain text, with good semantic headings.
Visual users will see the table. It is possible for users to
intentionally flip to the text-only view.