Added small-group to all pages.
Added the small group prompt include to most of the lessons. There's a
few where the structure was a bit different, so I did not. (And, in some
cases, I had to update the lesson to include a full-group, or explicitly
call out there not being a full group for the lesson.)

This served to both add the include to many of the pages, and do a bit
of "smoothing" across the content. Next, I need to cleanup layouts.