Cleaning up for print.
I wanted the rubric to print nicely. This gets pretty close.

There were some small changes to the rubric itself, and while I was at
it, I set the menu, header, and footer to be "no-print." This way, when
pages from the site are printed, they don't show up. This gives a clean
printing experience across the entire site.