Cleanup, new additions.
This is a big commit. I should have broken it up.

First, there are containers that are the start of work for automation
with Github Actions.

Second, there's cleanup related to making sure all of the lessons have
some kind of small group or full group discussion. They do. Any that
don't are because the lesson before "spills over" into the next lesson.

A "this is not official" banner has been added. This involved an include
and an addition to a template.

Work began on "related pages in the guides." This involved a new
datafile for managing the links out to the guides.