First pass at removing SO from SOMD
Did a pass through the entire site. References to state officers were
replaced with "product owner." In most cases (not all... :P) this was
done with a macro or two in the _config.yaml. If a different term wants
to be used, these macros might be enough to change the whole site.

A search for "product owner" should be made first, because I think the
first replace or two (lost in the mists of tabs...) were done by hand,
and then I quickly switched to a macro.